Implemented the rendering of cell values
Implemented the rendering of cell values. Theoretically the metaballs are working but with 2 - 10 fps. That is because the program does some heavy math for every cell. For the resolution of 960x540 that is 518400 cells. The plan is to use the marching squares algorithm to approximate the contour and thus optimize the program.